renamed files and moved a few more
