+++ src/modules/job-runner-coordination/job-runner/job-runner.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { JobRunnerService } from './job-runner.service';
 import { ClearCacheModule } from '../clear-cache/clear-cache.module';
 import { InitCacheModule } from '../init-cache/init-cache.module';
-import { LoadNextPostsModule } from '../../stage1-processing/load-next-posts/load-next-posts.module';
+import { LoadNextPostsModule } from '../load-next-posts/load-next-posts.module';
 import { PostRankerManagerModule } from '../post-ranker-manager/post-ranker-manager.module';
 
 @Module({
