+++ src/modules/job-runner-coordination/post-ranker-manager/post-ranker-manager.service.spec.ts
@@ -4,7 +4,7 @@ import { describe, expect, it, beforeEach } from '@jest/globals';
 import { QueryPoolService } from '../../stage1-processing/query-pool/query-pool.service';
 import { TlineCacherService } from '../../tline-cacher/tline-cacher.service';
 import { DispatcherService } from '../../stage2-processing/dispatcher/dispatcher.service';
-import { BatchProcessorService } from '../../stage2-processing/batch-processor/batch-processor.service';
+import { BatchProcessorService } from '../../stage3-processing/batch-processor/batch-processor.service';
 
 describe('PostRankerManagerService', () => {
     let service: PostRankerManagerService;
