+++ src/modules/job-runner-coordination/post-ranker-manager/post-ranker-manager.service.ts
@@ -3,7 +3,7 @@ import { JobResult, QueryJobListing } from '../../../utils/types';
 import { JobTypes } from '../../../utils/JobTypes';
 import { QueryPoolService } from '../../stage1-processing/query-pool/query-pool.service';
 import { DispatcherService } from '../../stage2-processing/dispatcher/dispatcher.service';
-import { BatchProcessorService } from '../../stage2-processing/batch-processor/batch-processor.service';
+import { BatchProcessorService } from '../../stage3-processing/batch-processor/batch-processor.service';
 import { TlineCacherService } from '../../tline-cacher/tline-cacher.service';
 
 @Injectable()
