+++ src/modules/stage1-processing/query-pool/query-pool.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { FeedQuerierModule } from '../feed-querier/feed-querier.module';
 import { QueryPoolService } from './query-pool.service';
-import { TLineCalculatorModule } from '../../stage2-processing/t-line-calculator/t-line-calculator.module';
+import { Stage2CalculationsModule } from '../../stage2-processing/stage2-calculations/stage2-calculations.module';
 
 @Module({
     providers: [QueryPoolService],
     exports: [QueryPoolService],
-    imports: [FeedQuerierModule, TLineCalculatorModule],
+    imports: [FeedQuerierModule, Stage2CalculationsModule],
 })
 export class QueryPoolModule {}
