+++ src/modules/stage1-processing/query-pool/query-pool.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { QueryPoolService } from './query-pool.service';
 import { describe, expect, it, beforeEach } from '@jest/globals';
-import { TLineCalculatorService } from '../../stage2-processing/t-line-calculator/t-line-calculator.service';
+import { Stage2CalculationsService } from '../../stage2-processing/stage2-calculations/stage2-calculations.service';
 import { FeedQuerier } from '../feed-querier/feed-querier.service';
 
 describe('QueryPoolService', () => {
@@ -18,7 +18,7 @@ describe('QueryPoolService', () => {
                     },
                 },
                 {
-                    provide: TLineCalculatorService,
+                    provide: Stage2CalculationsService,
                     useValue: {
                         calculateSectionsToQuery: jest.fn(),
                     },
