--- src/modules/stage2-processing/batch-processor/batch-processor.module.ts
@@ -1,11 +1,0 @@
-import { Module } from '@nestjs/common';
-import { BatchCalculatorService } from '../batch-calculator/batch-calculator.service';
-import { TlineCacherModule } from '../../tline-cacher/tline-cacher.module';
-import { TLineCalculatorModule } from '../t-line-calculator/t-line-calculator.module';
-
-@Module({
-    providers: [BatchCalculatorService],
-    exports: [BatchCalculatorService],
-    imports: [TlineCacherModule, TLineCalculatorModule],
-})
-export class BatchProcessorModule {}
