+++ src/modules/stage2-processing/dispatcher/dispatcher.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { DispatcherService } from './dispatcher.service';
-import { TLineCalculatorModule } from '../t-line-calculator/t-line-calculator.module';
+import { Stage2CalculationsModule } from '../stage2-calculations/stage2-calculations.module';
 import { BatchCalculatorModule } from '../batch-calculator/batch-calculator.module';
 
 @Module({
     providers: [DispatcherService],
     exports: [DispatcherService],
-    imports: [BatchCalculatorModule, TLineCalculatorModule],
+    imports: [BatchCalculatorModule, Stage2CalculationsModule],
 })
 export class DispatcherModule {}
