+++ src/modules/stage3-processing/batch-processor/batch-processor.service.ts
